% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{beta_matrix}
\alias{beta_matrix}
\title{Beta Matrix with Missing Values}
\format{
A matrix with 15 rows and 10 columns.
}
\description{
A 15 x 10 matrix of methylation beta values for sampled probes with some missing data.
(Probes are in rows and samples are in columns.)
}
\examples{
data(beta_matrix)
str(beta_matrix)
}
