% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methyl_summarize.R
\name{methyl_summarize}
\alias{methyl_summarize}
\title{Summarize Methylation Data}
\usage{
methyl_summarize(methyl_data, methyl_format)
}
\arguments{
\item{methyl_data}{A numeric matrix containing methylation data. Each row represents a probe, and each column represents a sample.}

\item{methyl_format}{A character string indicating the format of the methylation data. Must be either \code{"betas"} or \code{"mvals"}.}
}
\value{
A list containing two data frames:
\describe{
\item{\code{betas}}{A data frame summarizing the methylation data in beta values format.}
\item{\code{mvals}}{A data frame summarizing the methylation data in M-values format.}
}

Each data frame includes the following columns:
\itemize{
\item \code{probe}: The probe ID.
\item \code{mean}: Mean methylation value.
\item \code{sd}: Standard deviation.
\item \code{ci95_lb}: Lower bound of the 95\% confidence interval.
\item \code{ci95_ub}: Upper bound of the 95\% confidence interval.
\item \code{coef_var}: Coefficient of variation.
\item \code{iqr}: Interquartile range.
\item \code{min}: Minimum value.
\item \code{p05}: 5th percentile.
\item \code{p25}: 25th percentile (first quartile).
\item \code{p50}: Median (50th percentile).
\item \code{p75}: 75th percentile (third quartile).
\item \code{p95}: 95th percentile.
\item \code{max}: Maximum value.
\item \code{miss_n}: Number of missing values.
\item \code{nonmiss_n}: Number of non-missing values.
\item \code{miss_p}: Proportion of missing values.
\item \code{nonmiss_p}: Proportion of non-missing values.
}
}
\description{
The \code{methyl_summarize()} function provides a comprehensive summary of methylation data, either in beta values or M-values format. The function computes various statistics, including mean, standard deviation, confidence intervals, and more for each probe across samples.
}
\details{
The function checks whether the input \code{methyl_data} is a numeric matrix and whether \code{methyl_format} is valid. It then summarizes the data, providing extensive statistics for each probe. If the number of columns exceeds the number of rows in \code{methyl_data}, a warning is issued to ensure that the probes are organized as rows and samples as columns.
}
\examples{
\dontrun{
# example with betas as input
data(beta_matrix, package = "MethylSummarizR")
results <- methyl_summarize(methyl_data = beta_matrix, methyl_format = "betas")
# example with M-values as input
data(mval_matrix, package = "MethylSummarizR")
results <- methyl_summarize(methyl_data = mval_matrix, methyl_format = "mvals")
}

}
