% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{mval_matrix}
\alias{mval_matrix}
\title{M-Values Matrix with Missing Values}
\format{
A matrix with 15 rows and 10 columns.
}
\description{
A 15 x 10 matrix of methylation M-values without any missing data.
(Probes are in rows and samples are in columns.)
}
\examples{
data(mval_matrix)
str(mval_matrix)
}
